# adaptation-server/openapi.yaml
openapi: 3.0.3
info:
  title: Adaptation Admin API
  description: API for managing employee adaptation process
  version: 1.0.0
servers:
  - url: /api
    description: Default server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Error:
      type: object
      properties:
        message:
          type: string
          description: Error message
...
    Track:
      type: object
      required:
        - id
        - title
        - milestones
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        description:
          type: string
        milestones:
          type: array
          items:
            $ref: '#/components/schemas/Milestone'
    Milestone:
      type: object
      required:
        - id
        - title
        - description
        - steps
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        description:
          type: string
        steps:
          type: array
          items:
            oneOf:
              - $ref: '#/components/schemas/PresentationStep'
              - $ref: '#/components/schemas/TaskStep'
              - $ref: '#/components/schemas/SurveyStep'
    BaseStep:
      type: object
      required:
        - id
        - type
        - title
        - description
      properties:
        id:
          type: string
          format: uuid
        type:
          type: string
        title:
          type: string
        description:
          type: string
    PresentationStep:
      allOf:
        - $ref: '#/components/schemas/BaseStep'
        - type: object
          properties:
            type:
              type: string
              enum: [presentation]
            content:
              type: object
              properties:
                slides:
                  type: array
                  items:
                    oneOf:
                      - $ref: '#/components/schemas/ContentSlide'
                      - $ref: '#/components/schemas/QuizSlide'
    TaskStep:
      allOf:
        - $ref: '#/components/schemas/BaseStep'
        - type: object
          properties:
            type:
              type: string
              enum: [task]
            content:
              type: object
              properties:
                description:
                  type: string
                meeting:
                  type: object
                  properties:
                    title:
                      type: string
                    date:
                      type: string
                      format: date
                    time:
                      type: string
                    duration:
                      type: string
                    location:
                      type: string
                    participants:
                      type: array
                      items:
                        type: string
    SurveyStep:
      allOf:
        - $ref: '#/components/schemas/BaseStep'
        - type: object
          properties:
            type:
              type: string
              enum: [survey]
            content:
              type: object
              properties:
                title:
                  type: string
                description:
                  type: string
                form:
                  type: object
                  properties:
                    components:
                      type: array
                      items:
                        oneOf:
                          - $ref: '#/components/schemas/TextFieldComponent'
                          - $ref: '#/components/schemas/SelectComponent'
                          - $ref: '#/components/schemas/RatingComponent'
    ContentSlide:
      type: object
      properties:
        title:
          type: string
        content:
          type: string
    QuizSlide:
      type: object
      properties:
        type:
          type: string
          enum: [quiz]
        title:
          type: string
        content:
          type: object
          properties:
            question:
              type: string
            options:
              type: array
              items:
                type: string
            correctAnswer:
              type: integer
    BaseFormComponent:
      type: object
      required:
        - type
        - label
        - key
        - required
      properties:
        type:
          type: string
        label:
          type: string
        key:
          type: string
        required:
          type: boolean
    TextFieldComponent:
      allOf:
        - $ref: '#/components/schemas/BaseFormComponent'
        - type: object
          properties:
            type:
              type: string
              enum: [textfield]
            multiline:
              type: boolean
            placeholder:
              type: string
    SelectComponent:
      allOf:
        - $ref: '#/components/schemas/BaseFormComponent'
        - type: object
          properties:
            type:
              type: string
              enum: [select]
            options:
              type: array
              items:
                type: object
                properties:
                  label:
                    type: string
                  value:
                    type: string
    RatingComponent:
      allOf:
        - $ref: '#/components/schemas/BaseFormComponent'
        - type: object
          properties:
            type:
              type: string
              enum: [rating]
            min:
              type: integer
            max:
              type: integer
...
    Position:
      type: object
      required:
        - id
        - name
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
...
security:
  - bearerAuth: []
paths:
...
  /tracks:
    get:
      summary: Get all tracks
      responses:
        '200':
          description: List of tracks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Track'
    post:
      summary: Create a new track
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - title
                - milestones
              properties:
                title:
                  type: string
                description:
                  type: string
                milestones:
                  type: array
                  items:
                    $ref: '#/components/schemas/Milestone'
      responses:
        '201':
          description: Track created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Track'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /tracks/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    get:
      summary: Get track by ID
      responses:
        '200':
          description: Track details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Track'
        '404':
          description: Track not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    put:
      summary: Update track
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Track'
      responses:
        '200':
          description: Track updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Track'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Track not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      summary: Delete track
      responses:
        '204':
          description: Track deleted
        '404':
          description: Track not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /positions:
    get:
      summary: Get all positions
      responses:
        '200':
          description: List of positions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Position'
    post:
      summary: Create a new position
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
              properties:
                name:
                  type: string
                description:
                  type: string
      responses:
        '201':
          description: Position created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Position'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /positions/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    put:
      summary: Update position
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Position'
      responses:
        '200':
          description: Position updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Position'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Position not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      summary: Delete position
      responses:
        '204':
          description: Position deleted
        '404':
          description: Position not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...